Add state management functionality using redux.
Create a redux store and associated actions and reducers in order
to facilitate state management of the web app.

Also, get rid of a small bug which disallowed cors from working
properly wherein it was used into the application after the
postRoutes was declared. This order had to be reversed.