Add post creation capability and create form.
Allow users to create posts with the use of form that we have
created. At this point users can make posts and submit them,
which in turn creates POST request to the back end and shows
up the payload at the mongoDB database.

Get rid of a few typos, like replace all instances of 'createPosts'
with 'createPost'.