Add code for a post card and to update posts.
Add code for the styling of a post and all the buttons that appear
on all the posts. Also, add updatePost boilerplate code for the
backend which we shall later execute using the same form that we
designed earlier.

Add new background svg css code file which looks cool now.