Delete given file for deployment to heroku.
Initially, I figured heroku needed a package.json file to be in
the root directory of the project in order to be deployed correctly.

But since we have separate folder for server rather than its contents
being in the root folder, we would have to push in the contents of
only that folder in order for the deployment to work out fine.

Currently, while the push is doing alright, the app is ultimately
not working still after deployment. I just figured out the way to
push a specific folder to heroku needs the following command:

git subtree push --prefix <subfolder> heroku master

which is what we will try next.